Update ReportDocument.md
Moved a big block of text to section 4 - 'problem we are trying to solve' to improve flow 

While the MoCA framework is primarily a way of describing collaborative work, we propose that some aspects of the framework can also be used to describe collaborative software and systems. These similarities will be tested to see if well articulated CSCW can be matched to software tools suitable for the work through the notion of affordances. Users often maintain historically developed understandings of the utility of collaborative software and systems which are always evolving. Conole and Dyke (2004) argue that articulating and exposing the affordances of collaborative software enables users to understand how technologies can be most effectively used to support collaborative activities. So, we will also base our recommendation system on user input and perceptions.